Create game intro
This commit has a basic intro that first asks the user if they want to
play. Depending on the input, 'y' or 'n' the game is supposed to begin,
quit program or replay the intro. It also tells the user that Player 1's
marker is and X and Player 2's marker is an O.

I'm still thinking through how I want to construct the classes and how
they are supposed to be implemented. I think the bulk of the game itself
will be contained in the play_game() method but things may change as I
progress.