Class Board and play_game method
Class Board includes class variable, choices, to keep track of players'
marks and disable overlapping. The mark_board method within the class
uses the choices class variable and and if...else block to handle this.

The @@choice variable is an array that the player's take numbers from to
add to their respective arrays.

Within the play_game method, there are winning combination nested arrays
that the players can win by 'collecting'. Currently the game is set up
as a while loop (while game_won is false). The game_won variable is
reassigned to true once a player has collected a winning combination of
choices, regardless of order because the elements are sorted.

There are still lots of incomplete parts. The game doesn't properly end
even when a player has any of the three_in_a_row array combinations. The
game also doesn't properly reset the board when the user chooses to play
again.